requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
